Create two_factor_authentication_for_gmail.py
Implementing two-factor authentication (2FA) for email in Python typically involves sending a one-time code to the user's email or mobile device and then verifying that code when the user enters it. Below is an example of how you can implement 2FA for email using Python with the smtplib library for sending emails.

First, make sure you have a Gmail account or any other email service you want to use to send the verification code. You'll need to enable "Less secure apps" in your Gmail settings to use this approach. Be cautious about security when using this method in a production application.