Create gui_for_ip_address.py
creating a Python interface that obtains the IP address and MAC (Media Access Control) address of a user's computer, you can use the socket and uuid libraries. Here's a basic example of how to do this:

In this code:

The get_ip_address function creates a socket object and connects to a remote server (Google's public DNS server at 8.8.8.8). This connection is used to determine the local IP address of the user's computer.

The get_mac_address function uses the uuid library to obtain the MAC address of the user's computer.

Finally, we print the IP address and MAC address.

Please note that the ability to obtain the MAC address may be restricted in some environments or platforms due to security and privacy concerns. You may not be able to obtain the MAC address on all systems.